--- Cherprang.js
@@ -1,175 +1,0 @@
-//Get discord.js
-const Discord = require('discord.js');
-
-//require Other js file
-const token = require('./token.js');
-
-//request JSON form URL
-const request = require("request");
-const axios = require("axios");
-
-//Create client instance as bot
-const botCherprang = new Discord.Client();
-
-//Set listener on 'ready'
-botCherprang.on('ready', () => {
-  // console.log('Cherprang ready!');
-
-  /* status
-    - online
-    - idle
-    - invisible
-    - dnd
-  */
-  botCherprang.user.setStatus('online');
-
-  /* activity
-    type=0 = playing  
-    type=1 = streaming  
-    type=2 = listening 
-  */
-  botCherprang.user.setPresence({
-    game: {
-      name: 'BNK48 - BNK48',
-      type: 2
-    }
-  });
-
-});
-
-//Set listener on 'message'
-botCherprang.on('message', message => {
-
-  var messageRecsive = message.content;
-  var messageReply = '';
-
-  if (message.author.bot != true) {
-    if (messageRecsive.match(/^boardcast/)) {
-      message.delete();
-      messageReply = messageRecsive.replace('boardcast ', '');
-      message.channel.sendMessage('@everyone ' + messageReply);
-    } else if (messageRecsive.match(/^gif/)) {
-      var key = messageRecsive.replace('gif ', '');
-      key = key.replace(/ /g, "-");
-      var url = 'https://api.tenor.com/v1/random?limit=1&q=' + key;
-      axios.get(url)
-        .then(function (response) {
-          var img = response.data.results[0].itemurl
-          if (img == null || img == '') {
-            img = response.data.results[0].url
-          }
-          messageReply = 'นี่ค่ะ ' + img;
-          message.channel.sendMessage(messageReply);
-        })
-        .catch(function (error) {
-          messageReply = 'ไม่มีค่าาาาา XD';
-          message.channel.sendMessage(messageReply);
-        });
-    } else if (messageRecsive.match(/^gimme/)) {
-      var key = messageRecsive.replace('gimme ', '');
-      key = key.replace(/ /g, "-");
-      var url = 'https://api.tenor.com/v1/random?limit=1&q=' + key;
-      axios.get(url)
-        .then(function (response) {
-          var img = response.data.results[0].itemurl
-          if (img == null || img == '') {
-            img = response.data.results[0].url
-          }
-          messageReply = 'นี่ค่ะ ' + img;
-          message.channel.sendMessage(messageReply);
-        })
-        .catch(function (error) {
-          messageReply = 'ไม่มีค่าาาาา XD';
-          message.channel.sendMessage(messageReply);
-        });
-    } else if (messageRecsive.match(/สวัสดี/) || messageRecsive.match(/หวัดดี/) || messageRecsive.match(/hello/) || messageRecsive.match(/Hello/)) {
-      messageReply = 'สวัสดีค่าาาา';
-      message.reply(messageReply);
-    } else if (messageRecsive.match(/ฝันดี/) || messageRecsive.match(/ราตรีสวัส/) || messageRecsive.match(/good night/) || messageRecsive.match(/Good Night/) || messageRecsive.match(/goodnight/) || messageRecsive.match(/Goodnight/) || messageRecsive.match(/oyasumi/)) {
-      switch (random(3)) {
-        case 0:
-          messageReply = 'ฝันดีค่าาาา (*^ -^*)';
-          break;
-        case 1:
-          messageReply = 'ราตรีสวัสดิ์ค่าาาา (*^ -^*)';
-          break;
-        case 2:
-          messageReply = 'อย่าลืมฝันถึงเฌอน้าาาา (> 。<)';
-          break;
-      }
-      message.reply(messageReply);
-    } else if (messageRecsive.match(/เฌอปราง/) || messageRecsive.match(/เฌอ/) || messageRecsive.match(/แคปเฌอ/) || messageRecsive.match(/กับตัน/)) {
-      messageReply = 'เรียกหาเฌอเหรอคะ? (*^ -^*)';
-      message.reply(messageReply);
-    } else if (messageRecsive.match(/น่ารัก/)) {
-      messageReply = 'ขอบคุณค่าาาา (*^ -^*)';
-      message.reply(messageReply);
-    } else if (messageRecsive === 'Koisuru' || messageRecsive === 'koisuru') {
-      messageReply = 'Fortune Cookie♩♫♪';
-      message.channel.sendMessage(messageReply);
-    } else if (messageRecsive.match(/จริงเหรอ/)) {
-      messageReply = 'ใช่ที่สุดเลยยยย (> 。<)';
-      message.reply(messageReply);
-    } else if (messageRecsive === 'คุกกี้') {
-      messageReply = 'https://scontent.fbkk5-2.fna.fbcdn.net/v/t1.0-9/27751713_409511716139793_2905288473682211443_n.jpg?oh=b2082673dd01c183192b7d9c90bf3a76&oe=5B1460A2';
-      message.channel.sendMessage(messageReply);
-    } else if (messageRecsive === 'thinking') {
-      messageReply = 'https://i.redd.it/b7c67vkxjw6z.gif';
-      message.channel.sendMessage(messageReply);
-    } else if (messageRecsive.match(/อยากเล่นแมว/)) {
-      switch (random(2)) {
-        case 0:
-          messageReply = 'ไม่เอาน่า มาเล่นกับเฌอปรางดีกว่านะ >///<';
-          message.reply(messageReply);
-          break;
-        case 1:
-          request({
-            url: 'http://random.cat/meow.php',
-            json: true
-          }, function (error, response, body) {
-            messageReply = 'นี่ค่ะ ' + body.file;
-            message.channel.sendMessage(messageReply);
-          })
-          break;
-      }
-    } else if (messageRecsive.match(/แมว/)) {
-      request({
-        // url: 'http://random.cat/meow.php',
-        url: 'http://aws.random.cat/meow',
-        json: true
-      }, function (error, response, body) {
-        messageReply = 'นี่ค่ะ ' + body.file;
-        message.channel.sendMessage(messageReply);
-      })
-    } else if (messageRecsive.match(/แจม/)) {
-      messageReply = 'ตอนนี้แจมไม่อยู่ค่า คุยกับเฌอปรางก่อนละกันนะ (◠‿◠✿)';
-      message.reply(messageReply);
-    } else if (messageRecsive.match(/555\+/)) {
-      messageReply = 'https://tenor.com/view/cherprang-bnk48-gif-9849452';
-      message.channel.sendMessage(messageReply);
-    } else if (messageRecsive === 'ทำไรยุอะ') {
-      messageReply = 'แอบมองเธออยู่นะจ้ะ แต่เธอไม่รู้บ้างเยย';
-      message.reply(messageReply);
-    } else if (messageRecsive.match(/เหงา/)) {
-      messageReply = 'ไม่เป็นไร เดี๋ยวเฌออยู่เป็นเพื่อนเองน้า~ (◠‿◠✿)';
-      message.reply(messageReply);
-    } else if ((messageRecsive.match(/เดี๋ยว/) || messageRecsive.match(/เดว/)) && (messageRecsive.match(/มา/))) {
-      messageReply = 'เฌอจะรอน้า~ (◠‿◠✿)';
-      message.reply(messageReply);
-    } else if ((messageRecsive.match(/มา/)) && (messageRecsive.match(/ละ/))) {
-      messageReply = 'ยินดีต้อนรับค่า (*^ - ^*)♩♫♪';
-      message.reply(messageReply);
-    }
-
-    // console message
-    // if (message.author.bot == true) {
-    //   console.log('Bot : ' + messageReply);
-    // }
-  }
-});
-
-function random(num) {
-  return Math.floor((Math.random() * num));
-}
-
-botCherprang.login(token.sendToken());
