+++ commands/index.js
@@ -1,0 +1,7 @@
+const message = require('./message/message.js');
+
+module.exports = {
+    message: function (eventMessage) {
+        message.resMsg(eventMessage)
+    }
+}
