+++ common/random.js
@@ -1,0 +1,5 @@
+module.exports = {
+    random: function (num) {
+        return Math.floor((Math.random() * num));
+    }
+}
