+++ package.json
@@ -12,7 +12,7 @@
   "devDependencies": {},
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "node Cherprang.js"
+    "dev": "node bot.js"
   },
   "keywords": [
     "chargrang-bot"
